Added remote data fetching

+++ test/test_play_time.py
@@ -3,7 +3,7 @@ Test module for game time
 """
 import os
 from shutil import os as shos
-from grfc.game import play_time as gt, INPUT_FOLDER, GRFC_FILE
+from grfc.game import play_time as pt, INPUT_FOLDER, GRFC_FILE
 from . import ROUND
 
 
@@ -12,8 +12,8 @@ def test_file_available():
 
 
 def test_overall_time_offs():
-    data = [gt.time_for_players(GRFC_FILE, f'Round {round_nbr}') for round_nbr in range(1, 6)]
-    stats = gt.data_stats(*gt.valid_data(data))
+    data = [pt.time_for_players(f'Round {round_nbr}', GRFC_FILE) for round_nbr in range(1, 6)]
+    stats = pt.data_stats(*pt.valid_data(data))
     assert len(stats) == 4 and len(stats.columns) == 10
     assert 'turns in goals' in stats.index
     assert dict(stats.loc[:, 'Nicholas']) == {'matches played': 5.0,
@@ -23,22 +23,23 @@ def test_overall_time_offs():
 
 
 def test_generate_report():
-    report = gt.generate_report()
+    report = pt.generate_report(GRFC_FILE)
     assert report.startswith('<table ')
 
 
 def test_report_file():
-    shos.unlink('report.html')
-    gt.write_report(gt.generate_report())
+    if os.path.isfile('report.html'):
+        shos.unlink('report.html')
+    pt.write_report(pt.generate_report(GRFC_FILE))
     assert os.path.isfile('report.html')
 
 
 def test_time_for_players():
-    times = gt.time_for_players(GRFC_FILE, ROUND)
+    times = pt.time_for_players(ROUND, GRFC_FILE)
     assert times[0]['Angus'] == 34.0
 
 
 def test_goalies_stats():
     goalies_list = ['Nicholas', 'Herik', 'Nicholas', 'Noah']
-    stats = gt.goalies_stats(goalies_list)
+    stats = pt.goalies_stats(goalies_list)
     assert 'Nicholas' in stats
